added roles, seo tags, contact us page

+++ routes/web.php
@@ -28,7 +28,7 @@ use App\Models\SocialSetting;
 // Static Pages
 Route::get('/', [StaticPagesController::class, 'home']);
 Route::get('/menu', [StaticPagesController::class, 'menu']);
-Route::get('/menu/id', [StaticPagesController::class, 'singleMenu']);
+Route::get('/menu/{slug}', [StaticPagesController::class, 'singleMenu']);
 Route::get('/about', [StaticPagesController::class, 'about']);
 Route::get('/reservations', [StaticPagesController::class, 'reservations']);
 Route::post('/reservations', [StaticPagesController::class, 'saveReservation']);
@@ -43,20 +43,20 @@ Route::get('/offers/thank-you', [StaticPagesController::class, 'thankYou']);
 Route::get('/admin', [AdminController::class, 'dashboard']);
 
 // Admin Food Categories
-Route::get('/admin/food-categories', [FoodCategoriesController::class, 'index']);
-Route::get('/admin/food-categories/create', [FoodCategoriesController::class, 'create']);
-Route::post('/admin/food-categories', [FoodCategoriesController::class, 'store']);
-Route::get('/admin/food-categories/{id}/edit', [FoodCategoriesController::class, 'edit']);
-Route::put('/admin/food-categories/{id}', [FoodCategoriesController::class, 'update']);
-Route::delete('/admin/food-categories/{id}/delete', [FoodCategoriesController::class, 'delete']);
+Route::get('/admin/food-categories', [FoodCategoriesController::class, 'index'])->middleware('role:Admin');
+Route::get('/admin/food-categories/create', [FoodCategoriesController::class, 'create'])->middleware('role:Admin');
+Route::post('/admin/food-categories', [FoodCategoriesController::class, 'store'])->middleware('role:Admin');
+Route::get('/admin/food-categories/{id}/edit', [FoodCategoriesController::class, 'edit'])->middleware('role:Admin');
+Route::put('/admin/food-categories/{id}', [FoodCategoriesController::class, 'update'])->middleware('role:Admin');
+Route::delete('/admin/food-categories/{id}/delete', [FoodCategoriesController::class, 'delete'])->middleware('role:Admin');
 
 // Admin Food Items
-Route::get('/admin/food-items', [FoodItemsController::class, 'index']);
-Route::get('/admin/food-items/create', [FoodItemsController::class, 'create']);
-Route::post('/admin/food-items', [FoodItemsController::class, 'store']);
-Route::get('/admin/food-items/{id}/edit', [FoodItemsController::class, 'edit']);
-Route::put('/admin/food-items/{id}', [FoodItemsController::class, 'update']);
-Route::delete('/admin/food-items/{id}/delete', [FoodItemsController::class, 'delete']);
+Route::get('/admin/food-items', [FoodItemsController::class, 'index'])->middleware('role:Admin');
+Route::get('/admin/food-items/create', [FoodItemsController::class, 'create'])->middleware('role:Admin');
+Route::post('/admin/food-items', [FoodItemsController::class, 'store'])->middleware('role:Admin');
+Route::get('/admin/food-items/{id}/edit', [FoodItemsController::class, 'edit'])->middleware('role:Admin');
+Route::put('/admin/food-items/{id}', [FoodItemsController::class, 'update'])->middleware('role:Admin');
+Route::delete('/admin/food-items/{id}/delete', [FoodItemsController::class, 'delete'])->middleware('role:Admin');
 
 // Admin Customers
 // Admin Members
@@ -68,21 +68,21 @@ Route::get('/admin/reservations', [ReservationController::class, 'index']);
 Route::delete('/admin/reservations/{id}/delete', [ReservationController::class, 'delete']);
 
 // Admin Settings
-Route::get('/admin/settings/general', [SettingController::class, 'general']);
-Route::put('/admin/settings/general', [SettingController::class, 'saveGeneral']);
-Route::get('/admin/settings/seo', [SettingController::class, 'seo']);
-Route::put('/admin/settings/seo', [SettingController::class, 'saveSeo']);
-Route::get('/admin/settings/social', [SettingController::class, 'social']);
-Route::put('/admin/settings/social', [SettingController::class, 'saveSocial']);
+Route::get('/admin/settings/general', [SettingController::class, 'general'])->middleware('role:Admin');
+Route::put('/admin/settings/general', [SettingController::class, 'saveGeneral'])->middleware('role:Admin');
+Route::get('/admin/settings/seo', [SettingController::class, 'seo'])->middleware('role:Admin');
+Route::put('/admin/settings/seo', [SettingController::class, 'saveSeo'])->middleware('role:Admin');
+Route::get('/admin/settings/social', [SettingController::class, 'social'])->middleware('role:Admin');
+Route::put('/admin/settings/social', [SettingController::class, 'saveSocial'])->middleware('role:Admin');
 
 
 // Admin Users
-Route::get('/admin/users', [UsersController::class, 'index']);
-Route::get('/admin/users/create', [UsersController::class, 'create']);
-Route::post('/admin/users', [UsersController::class, 'store']);
-Route::get('/admin/users/{id}/edit', [UsersController::class, 'edit']);
-Route::put('/admin/users/{id}', [UsersController::class, 'update']);
-Route::delete('/admin/users/{id}/delete', [UsersController::class, 'delete']);
+Route::get('/admin/users', [UsersController::class, 'index'])->middleware('role:Admin');
+Route::get('/admin/users/create', [UsersController::class, 'create'])->middleware('role:Admin');
+Route::post('/admin/users', [UsersController::class, 'store'])->middleware('role:Admin');
+Route::get('/admin/users/{id}/edit', [UsersController::class, 'edit'])->middleware('role:Admin');
+Route::put('/admin/users/{id}', [UsersController::class, 'update'])->middleware('role:Admin');
+Route::delete('/admin/users/{id}/delete', [UsersController::class, 'delete'])->middleware('role:Admin');
 
 // Admin Authentication
 Route::get('/admin/register', function () {
@@ -95,7 +95,7 @@ Route::get('/admin/login', function () {
 Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-View::composer(['home', 'pages.about', 'pages.contact', 'pages.offers', 'pages.reservations', 'thank-you', 'menu.index', 'menu.single-menu'], function($view) {
+View::composer(['home', 'pages.about', 'pages.contact', 'pages.offers', 'pages.reservations', 'pages.thank-you', 'menu.index', 'menu.single-menu'], function($view) {
     $general_settings = GeneralSetting::find(1);
     $seo_settings = SeoSetting::find(1);
     $social_settings = SocialSetting::find(1);
